pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = ""
authors = [
    {name = "Giles Strong", email = "giles.c.strong@gmail.com"},
]
readme = "README.md"
keywords = []
requires-python = "~=3.12.0"
dependencies = [
  "django>=6.0.3",
  "djangorestframework>=3.15.2",
  "drf-yasg>=1.21.8",
  "django-ninja>=1.3.0",
  "psycopg[binary]>=3.2.5",
  "python-dotenv>=1.0.1",
  "celery>=5.4.0",
  "redis>=5.2.1",
  "django-celery-beat>=2.7.0",
  "beartype>=0.22.9",
  "langchain-community>=0.4.1",
  "jq>=1.11.0",
  "tqdm>=4.67.1",
  "kombu>=5.6.2",
  "pydantic-ai>=1.58.0",
  "pydantic>=2.12.5",
  "pydantic-settings>=2.12.0",
  "qdrant-client>=1.16.2",
  "requests>=2.32.5",
  "logfire>=4.22.0",
  "aiocache>=0.12.3",
  "hypercorn>=0.18.0",
  "nanoid>=2.0.0",
  "google-auth-oauthlib>=1.2.4",
  "google-auth>=2.48.0",
  "pyjwt>=2.11.0",
  "cryptography>=46.0.5",
]

[dependency-groups]
dev = [
  "pre-commit>=3.6.1",
  "mypy>=1.8.0",
  "pytest-mock>=3.12.0",
  "flaky>=3.7.0",
  "pytest-lazy-fixture>=0.6.3",
  "pytest<8.0.0",
  "ruff==0.14.14",
  "pytest-asyncio>=0.23.8",
  "types-tqdm>=4.67.3.20260205",
]

[build-system]
requires = ["hatchling>=1.25"] 
build-backend = "hatchling.build"

[tool.ruff]
line-length = 120
target-version = "py312"
extend-exclude = [
  "dev/",
  "**/*.ipynb",
]

[tool.ruff.lint]
select = ["E", "F", "W", "B", "C90", "I"]
ignore = ["B007", "E203", "E266", "E501", "E741", "C901", "E402", "E211", "W291", "F722", "B904"]

[tool.ruff.lint.mccabe]
max-complexity = 18

[tool.ruff.format]
quote-style = "preserve"

[tool.mypy]
python_version = "3.12"
disallow_untyped_calls = false
disallow_untyped_defs = true
check_untyped_defs = true
exclude = "tests/|\\.venv/|dev/|migrations/"
no_strict_optional = true
disable_error_code = "attr-defined, override, union-attr"

[[tool.mypy.overrides]]
module = [
  "django.*",
  "celery.*",
  "kombu.*",
  "aiocache.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
addopts = "-m 'not long'"
asyncio_mode = "auto"

[tool.uv]
package = false
